Inital upload to Github
These routines provide a way for printing some information during the execution of an assembler programme. It's primary intended for usage with a serial interface (based on 16C2552), but it's easy to adapt the routines to stream the output to the screen or a printer.
Please note that you have to take care for saving important registers. Routines like debug_hex will destroy the contents in A, so if you need that information then you have to push and pull the A register